requirements_gpu.txt
# open-clip-torch==2.7.0
# fvcore
# cython
# albumentations
# boto3
# checksumdir
# clip
# decord
# easydict
# einops
# fairscale
# ftfy
# jieba
# kornia
# matplotlib
# more-itertools
# ninja
# opencv-python
# openpyxl==3.1.1
# omegaconf
# psutil
# PyYAML
# pywavelets
# pycocotools
# pytorch_lightning
# py-spy
# regex
# scipy
# sentencepiece
# streamlit
# tqdm
# tensorboard
# taming-transformers-rom1504
# transformers==4.30.2
# timm
# webdataset
# pandas
# deepspeed==0.12.6
# diffusers
# accelerate
# rouge-metric
# safetensors
tzlocal==4.3
grpcio==1.41.1
decorator  
basicsr
ftfy
regex
fairscale
jieba
einops
transformers
tb-nightly==2.12.0a20230113
clip
kornia
streamlit
py-spy
sentencepiece
easydict
boto3
checksumdir
jieba
omegaconf
matplotlib
tensorboard
opencv-python
scipy
albumentations
openpyxl
fvcore
cython
more-itertools
ninja
psutil
PyYAML
pywavelets
pycocotools
py-spy
tqdm
timm==0.9.12
webdataset
deepspeed==0.12.6
pandas<2.0.0
diffusers==0.21.4
onnxruntime
rouge_metric
pathlib
open-clip-torch==2.7.0
lpips